Map objects rendering correctly. Reports added below map

+++ backendathon/migrations/20260831200149_create_reports_table.js
@@ -8,8 +8,8 @@ exports.up = function (knex) {
     table.string("first_name", 100).notNullable();
     table.string("last_name", 100);
     table.string("user_grade_and_rank", 100);
-    table.integer("report_lat");
-    table.integer("report_long");
+    table.float("report_lat");
+    table.float("report_long");
     table.string("syscap_status_code", 250);
     table.string("response_actions", 250);
     table.string("comments", 250);
